requirements.txt
# Core dependencies
requests>=2.31.0
matplotlib>=3.8.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Telegram integration
python-telegram-bot>=20.7

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.12.0
# mypy>=1.7.0
# ruff>=0.1.8
